+++ src/main/scala/org/economicsl/auctions/multiunit/PriceQuantitySchedule.scala
@@ -13,7 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions
+package org.economicsl.auctions.multiunit
+
+import org.economicsl.auctions.{Order, Price, Quantity, Tradable}
 
 import scala.collection.GenIterable
